250321 1438

+++ src/main/java/com/app/pickcourse/domain/dto/ParticipantDTO.java
@@ -6,15 +6,13 @@ import org.springframework.stereotype.Component;
 
 @Component
 @NoArgsConstructor
-@Getter
-@Setter
-@ToString
+@Getter @Setter @ToString
 @EqualsAndHashCode(onlyExplicitlyIncluded = true)
 public class ParticipantDTO {
     @EqualsAndHashCode.Include
     private Long id;
-    private int memberId;
-    private int planId;
+    private Long memberId;
+    private Long planId;
     private String createdDate;
     private String updatedDate;
